cus增加介面跟實作
